Fix actions - final commit for the time being
Unfortunately we realized that this gym_tetris implementation is not
good for our course project goals. This gym environmentis designed
to work with an extremely deep network agent that reads a screen
and submits actions to the emulator.

When we submit an action to the emulator, (i.e. pressing the right
button to move the current falling Tetrimino piece to the right), we
would usually expect that the piece will move to the right. However,
since environment is merely a direct interface to the NES Tetris
emulator, the piece may not always move due to reasons like
"Delayed Auto Shift" (DAS).

This environment is not suited for direct comparisons with most of
the established Tetris RL literature which utilizes a very
simplified environment, and for good reason.

I will probably be making this a summer project. I think I will
fork gym_tetris and make it well-equipped with features for simpler
RL development.